+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased v0.7]
 
+## 2022-03-22/kw
+# changed
+- response after storing new offer returns complete offer with related data (competences etc)
+
 ## 2022-03-18/ml
 # changed
 - MiniDataSet (id, title, image, visible + sort_flag), is sorted bei sort_flag